move ChangeSpriteOnHover.cs to UI scripts
