+++ Assets/Scripts/UI/ChangeSpriteOnHover.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dca3c86d1109b674da912fb83b8d356e
+guid: 49937f7440bf0784aa0237893443060d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
